Fix: Gaussian distribution of weight matrix
The std deviation (`l`) is apparently supposed to be dependent on the
distance between neighbouring spots in visium data. Because with a fixed
value, which we've chosen to be very small (0.5), the permutation tests
would virtually give the same result as the actual data, because the
neighbours of a spot wouldn't have any effect on the max eigenvalue of a
spot.